+++ src/Symfony/Component/Tui/CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
  * Add `AbstractWidget::attachChild()` and `AbstractWidget::detachChild()` to wire child widgets
  * Add multi-select support to `SelectListWidget`
  * [BC BREAK] Add `$multiselect` as the third argument of `SelectListWidget::__construct()`, moving `$keybindings` to fourth position
+ * Add `KeyBindingWidget` to display the keybindings of the focused widget
 
 8.1
 ---